requirements.txt
flask
flask-script
flask-sqlalchemy
flask-restful
psycopg2
requests
beautifulsoup4
